feat(DeviceInfoObjectIter): Add a DeviceInfoObject iterator.
For a better API experience i added a first version of a Iterator that contains the received InfoObject bytes from the server and can give one object at a time to the user. There are still some issues and open Questions in regards how it should work so this is more of a experiment at the moment and not the final implementation.